Move v2.0 taggings controller into its own file
